Generic types to generic
